+++ scout/adapter/mongo/variant_loader.py
@@ -19,6 +19,7 @@ from intervaltree import IntervalTree
 from scout.parse.variant.headers import (parse_rank_results_header,
                                          parse_vep_header)
 from scout.parse.variant.rank_score import parse_rank_score
+from scout.parse.variant.clnsig import is_pathogenic
 
 from scout.parse.variant import parse_variant
 from scout.build import build_variant
@@ -389,12 +390,13 @@ class VariantLoader(object):
             # All MT variants are loaded
             mt_variant = 'MT' in variant.CHROM
             rank_score = parse_rank_score(variant.INFO.get('RankScore'), case_obj['_id'])
+            pathogenic = is_pathogenic(variant)
 
             # Check if the variant should be loaded at all
             # if rank score is None means there are no rank scores annotated, all variants will be loaded
             # Otherwise we load all variants above a rank score treshold
             # Except for MT variants where we load all variants
-            if (rank_score is None) or (rank_score > rank_threshold) or mt_variant:
+            if (rank_score is None) or (rank_score > rank_threshold) or mt_variant or pathogenic:
                 nr_inserted += 1
                 # Parse the vcf variant
                 parsed_variant = parse_variant(