+++ CHANGELOG.md
@@ -12,6 +12,9 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - New looks for Scout
 - Improved dashboard using Chart.js
 - Ask before inactivating a case where last assigned user leaves it
+- Genes can be manually added to the dynamic gene list directly on the case page
+- Dynanmic gene panels can optionally be used with clinical filter, instead of default gene panel
+- Load all clinvar variants with clinvar Pathogenic, Likely Pathogenic and Conflicting pathogenic
 
 ### fixed
 - Fixed missing import for variants with comments